requirements.txt
flask==1.1.2
flask_restful==0.3.8
flask_script==2.0.6
flask_migrate==2.1.1
marshmallow==3.5.1
flask_sqlalchemy==2.3.2
flask_marshmallow==0.8.0
marshmallow-sqlalchemy>=0.13.2
flask-restplus>=0.13.0
psycopg2>=2.7.5
numpy>=1.16.3
pymongo>=3.8.0
Werkzeug==0.16.1
more_itertools>=8.2.0
flask-mongoengine>=0.9.5
xlrd>=1.0.0
flask-excel>=0.0.7


waitress~=2.1.2
pythonnet>=2.3.0
tqdm~=4.66.1
pandas==1.4.1
jinja2==3.0.3
itsdangerous==2.0.1
numpy
matplotlib~=3.8.0
seaborn~=0.13.0

fastapi==0.83.0
uvicorn>=0.18.3
sqlalchemy>=1.4.41
bcrypt>=4.0.0
pydantic~=1.10.2
pydantic[email]
python-dotenv~=0.21.0
starlette>=0.19.1
python-multipart>=0.0.6
openpyxl>=3.1.2
mongoengine~=0.27.0
requests~=2.31.0